@@ -1,127 +1,12 @@
 {
-  "name": "challenges",
+  "name": "osumbchallenges",
   "version": "1.0.0",
-  "author": "Alex Tareshawty",
-  "description": "Challenge signup manager for OSUMB",
-  "engines": {
-    "node": "6.11.1"
-  },
-  "private": true,
+  "description": "Manage the challenge process for the Ohio State University Marching Band",
   "scripts": {
-    "start": "node bin/www",
-    "build": "node client/scripts/build.js",
-    "dev": "heroku local -f Procfile.dev",
-    "lint": "eslint .",
-    "heroku-postbuild": "npm run build",
-    "schema": "script/schema",
-    "seed": "./bin/seed",
-    "server": "nodemon --ignore client/ bin/www",
-    "test": "bin/test"
+    "heroku-prebuild": "bin/client_install",
+    "heroku-postbuild": "bin/client_build"
   },
-  "dependencies": {
-    "async": "^2.0.0-rc.5",
-    "autoprefixer": "6.5.1",
-    "autoprefixer-loader": "^3.2.0",
-    "babel-core": "6.17.0",
-    "babel-eslint": "7.0.0",
-    "babel-jest": "17.0.2",
-    "babel-loader": "6.2.7",
-    "babel-preset-react-app": "^2.0.0",
-    "bcrypt": "^0.8.6",
-    "body-parser": "~1.13.2",
-    "cancelable-promise": "^2.2.0",
-    "case-sensitive-paths-webpack-plugin": "1.1.4",
-    "chalk": "1.1.3",
-    "change-case": "^3.0.0",
-    "classnames": "^2.2.5",
-    "connect-history-api-fallback": "1.3.0",
-    "connect-redis": "^3.0.2",
-    "cookie-parser": "~1.3.5",
-    "cross-spawn": "4.0.2",
-    "css-loader": "0.26.0",
-    "db-migrate": "^0.10.0-beta.20",
-    "db-migrate-pg": "^0.1.11",
-    "debug": "~2.2.0",
-    "detect-port": "1.0.1",
-    "dotenv": "^2.0.0",
-    "es6-promise": "^3.2.1",
-    "express": "~4.13.1",
-    "express-handlebars": "^3.0.0",
-    "express-session": "^1.13.0",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "0.9.0",
-    "filesize": "3.3.0",
-    "fs-extra": "0.30.0",
-    "github-api": "^2.3.0",
-    "gzip-size": "3.0.0",
-    "html-webpack-plugin": "2.24.0",
-    "http-proxy-middleware": "0.17.2",
-    "jest": "17.0.2",
-    "json-loader": "0.5.4",
-    "jsonwebtoken": "^7.2.0",
-    "jwt-decode": "^2.1.0",
-    "keycode": "^2.1.7",
-    "material-ui": "^0.16.4",
-    "moment": "^2.12.0",
-    "morgan": "^1.7.0",
-    "node-cron": "^1.1.2",
-    "node-sass": "^3.13.0",
-    "node-xlsx": "^0.7.1",
-    "nodemon": "^1.10.0",
-    "object-assign": "4.1.0",
-    "passport": "^0.3.2",
-    "passport-local": "^1.0.0",
-    "path-exists": "2.1.0",
-    "pg": "^4.5.1",
-    "pg-pool": "^1.4.3",
-    "postcss-loader": "^1.0.0",
-    "promise": "7.1.1",
-    "react": "^15.4.1",
-    "react-datetime": "^2.8.1",
-    "react-dev-utils": "^0.4.0",
-    "react-dom": "^15.4.1",
-    "react-media": "^1.4.0",
-    "react-router-dom": "4.0.0-beta.3",
-    "react-tap-event-plugin": "^2.0.1",
-    "recursive-readdir": "2.1.0",
-    "redis": "^2.6.0-2",
-    "sass-loader": "^3.2.3",
-    "sendgrid": "^3.0.5",
-    "serve-favicon": "~2.3.0",
-    "strip-ansi": "3.0.1",
-    "style-loader": "0.13.1",
-    "url-loader": "0.5.7",
-    "validator": "^6.2.1",
-    "webpack": "^1.13.1",
-    "webpack-dev-server": "1.16.2",
-    "webpack-manifest-plugin": "1.1.0",
-    "whatwg-fetch": "^1.0.0"
-  },
-  "devDependencies": {
-    "@btmills/eslint-config-btmills": "^1.0.0-beta.6",
-    "eslint": "3.8.1",
-    "eslint-plugin-react": "^6.3.0",
-    "jasmine": "^2.4.1"
-  },
-  "jest": {
-    "collectCoverageFrom": [
-      "src/**/*.{js,jsx}"
-    ],
-    "moduleNameMapper": {
-      "^.+\\.(ico|jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/config/jest/FileStub.js",
-      "^.+\\.css$": "<rootDir>/config/jest/CSSStub.js"
-    },
-    "setupFiles": [
-      "<rootDir>/config/polyfills.js"
-    ],
-    "testPathIgnorePatterns": [
-      "<rootDir>/(build|docs|node_modules)/"
-    ],
-    "testEnvironment": "node"
-  },
-  "babel": {
-    "presets": [
-      "react-app"
-    ]
+  "engines": {
+    "node": "7.7.1"
   }
 }
