@@ -1,21 +1,17 @@
-/* eslint-disable global-require */
+/* eslint-disable global-require, react/jsx-filename-extension */
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
-import './index.scss';
 
-ReactDOM.render(
-  <App />,
-  document.getElementById('root')
-);
+import './index.css';
+import 'material-components-web/dist/material-components-web.min.css';
+import App from './App';
+
+ReactDOM.render(<App />, document.getElementById('root'));
 
 if (module.hot) {
   module.hot.accept('./App', () => {
     const NextApp = require('./App').default;
 
-    ReactDOM.render(
-      <NextApp />,
-      document.getElementById('root')
-    );
+    ReactDOM.render(<NextApp />, document.getElementById('root'));
   });
 }
