@@ -1,4 +1,4 @@
-import React, { Component, PropTypes } from 'react';
+import React, { Component } from 'react';
 import { Card, CardActions, CardHeader, CardText } from 'material-ui/Card';
 import Dialog from 'material-ui/Dialog';
 import Divider from 'material-ui/Divider';
@@ -6,6 +6,7 @@ import FlatButton from 'material-ui/FlatButton';
 import { grey500 } from 'material-ui/styles/colors';
 import LockOpen from 'material-ui/svg-icons/action/lock-open';
 import LockOutline from 'material-ui/svg-icons/action/lock-outline';
+import PropTypes from 'prop-types';
 import { RadioButton, RadioButtonGroup } from 'material-ui/RadioButton';
 import Snackbar from 'material-ui/Snackbar';
 import TextField from 'material-ui/TextField';
@@ -14,7 +15,6 @@ import './pending-result.scss';
 import { api } from '../utils';
 
 export default class PendingResult extends Component {
-
   static get propTypes() {
     return {
       firstComments: PropTypes.string.isRequired,
@@ -59,7 +59,11 @@ export default class PendingResult extends Component {
   }
 
   handleCancel() {
-    const { originalFirstComments, originalSecondComments, originalWinnerId } = this.state;
+    const {
+      originalFirstComments,
+      originalSecondComments,
+      originalWinnerId
+    } = this.state;
 
     this.setState({
       editing: false,
@@ -88,22 +92,23 @@ export default class PendingResult extends Component {
   handleSubmit() {
     const { firstComments, secondComments, winnerId } = this.state;
 
-    api.put('/results/pending', {
-      id: this.props.id,
-      firstComments,
-      secondComments,
-      winnerId
-    })
-    .then(() => {
-      this.setState({
-        confirming: false,
-        editing: false,
-        originalFirstComments: firstComments,
-        originalSecondComments: secondComments,
-        originalWinnerId: winnerId,
-        success: true
+    api
+      .put('/results/pending', {
+        id: this.props.id,
+        firstComments,
+        secondComments,
+        winnerId
+      })
+      .then(() => {
+        this.setState({
+          confirming: false,
+          editing: false,
+          originalFirstComments: firstComments,
+          originalSecondComments: secondComments,
+          originalWinnerId: winnerId,
+          success: true
+        });
       });
-    });
   }
 
   handleWinnerIdChange(event, value) {
@@ -112,7 +117,14 @@ export default class PendingResult extends Component {
     });
   }
 
-  renderRadioButtons(firstName, firstNameNumber, id, secondName, secondNameNumber, winnerId) {
+  renderRadioButtons(
+    firstName,
+    firstNameNumber,
+    id,
+    secondName,
+    secondNameNumber,
+    winnerId
+  ) {
     if (secondName && secondNameNumber) {
       return (
         <RadioButtonGroup
@@ -120,14 +132,8 @@ export default class PendingResult extends Component {
           onChange={this.handleWinnerIdChange}
           valueSelected={winnerId}
         >
-          <RadioButton
-            label={firstName}
-            value={firstNameNumber}
-          />
-          <RadioButton
-            label={secondName}
-            value={secondNameNumber}
-          />
+          <RadioButton label={firstName} value={firstNameNumber} />
+          <RadioButton label={secondName} value={secondNameNumber} />
         </RadioButtonGroup>
       );
     } else {
@@ -137,34 +143,49 @@ export default class PendingResult extends Component {
           onChange={this.handleWinnerIdChange}
           valueSelected={winnerId}
         >
-          <RadioButton
-            label={firstName}
-            value={firstNameNumber}
-          />
+          <RadioButton label={firstName} value={firstNameNumber} />
         </RadioButtonGroup>
       );
     }
-
   }
 
   render() {
-    const { confirming, editing, firstComments, secondComments, success, winnerId } = this.state;
-    const { firstName, firstNameNumber, id, secondName, secondNameNumber, spotId } = this.props;
+    const {
+      confirming,
+      editing,
+      firstComments,
+      secondComments,
+      success,
+      winnerId
+    } = this.state;
+    const {
+      firstName,
+      firstNameNumber,
+      id,
+      secondName,
+      secondNameNumber,
+      spotId
+    } = this.props;
     const currentWinner = winnerId === firstNameNumber ? firstName : secondName;
     const dialogActions = [
-      <FlatButton key="cancel" onTouchTap={this.handleConfirmClose}>No</FlatButton>,
-      <FlatButton key="submit" onTouchTap={this.handleSubmit}>Submit</FlatButton>
+      <FlatButton key="cancel" onTouchTap={this.handleConfirmClose}>
+        No
+      </FlatButton>,
+      <FlatButton key="submit" onTouchTap={this.handleSubmit}>
+        Submit
+      </FlatButton>
     ];
     const title = editing
-      ? <strong><LockOpen /> {currentWinner} ({spotId})</strong>
-      : <strong><LockOutline /> {currentWinner} ({spotId})</strong>;
+      ? <strong>
+          <LockOpen /> {currentWinner} ({spotId})
+        </strong>
+      : <strong>
+          <LockOutline /> {currentWinner} ({spotId})
+        </strong>;
 
     return (
       <div className="PendingResult">
-        <Dialog
-          actions={dialogActions}
-          open={confirming}
-        >
+        <Dialog actions={dialogActions} open={confirming}>
           Are you sure you want to update this result&#63;
         </Dialog>
         <Snackbar
@@ -180,45 +201,83 @@ export default class PendingResult extends Component {
             }}
             title={title}
           />
-          {editing ?
-            <div className="PendingResult-textHeaders">
-              {this.renderRadioButtons(firstName, firstNameNumber, id, secondName, secondNameNumber, winnerId)}
-            </div> :
-            <div className="PendingResult-textHeaders">
-              <p>{firstName}</p>
-              {secondNameNumber && <p>{secondName}</p>}
-            </div>
-          }
+          {editing
+            ? <div className="PendingResult-textHeaders">
+                {this.renderRadioButtons(
+                  firstName,
+                  firstNameNumber,
+                  id,
+                  secondName,
+                  secondNameNumber,
+                  winnerId
+                )}
+              </div>
+            : <div className="PendingResult-textHeaders">
+                <p>
+                  {firstName}
+                </p>
+                {secondNameNumber &&
+                  <p>
+                    {secondName}
+                  </p>}
+              </div>}
           <Divider />
           <div className="PendingResult-text">
-            {editing ?
-              <TextField
-                multiLine
-                name="firstComments"
-                onChange={this.handleCommentsChange}
-                rows={4}
-                rowsMax={4}
-                value={firstComments}
-              /> :
-              <CardText>{firstComments}</CardText>
-            }
-            {secondNameNumber && (editing ?
-              <TextField
-                multiLine
-                name="secondComments"
-                onChange={this.handleCommentsChange}
-                rows={4}
-                rowsMax={4}
-                value={secondComments}
-              /> :
-              <CardText>{secondComments}</CardText>
-            )}
+            {editing
+              ? <TextField
+                  multiLine
+                  name="firstComments"
+                  onChange={this.handleCommentsChange}
+                  rows={4}
+                  rowsMax={4}
+                  value={firstComments}
+                />
+              : <CardText>
+                  {firstComments}
+                </CardText>}
+            {secondNameNumber &&
+              (editing
+                ? <TextField
+                    multiLine
+                    name="secondComments"
+                    onChange={this.handleCommentsChange}
+                    rows={4}
+                    rowsMax={4}
+                    value={secondComments}
+                  />
+                : <CardText>
+                    {secondComments}
+                  </CardText>)}
           </div>
           <CardActions>
-            {editing && <FlatButton className="PendingResult-button" onTouchTap={this.handleCancel}>Cancel</FlatButton>}
-            {editing && <FlatButton className="PendingResult-button" onTouchTap={this.handleConfirm}>Update</FlatButton>}
-            {!editing && <FlatButton className="PendingResult-button" onTouchTap={this.handleApproveClick}>Approve</FlatButton>}
-            {!editing && <FlatButton className="PendingResult-button" onTouchTap={this.handleEditClick}>Edit</FlatButton>}
+            {editing &&
+              <FlatButton
+                className="PendingResult-button"
+                onTouchTap={this.handleCancel}
+              >
+                Cancel
+              </FlatButton>}
+            {editing &&
+              <FlatButton
+                className="PendingResult-button"
+                onTouchTap={this.handleConfirm}
+              >
+                Update
+              </FlatButton>}
+            {!editing &&
+              <FlatButton
+                className="PendingResult-button"
+                onTouchTap={this.handleApproveClick}
+              >
+                Approve
+              </FlatButton>}
+            {!editing &&
+              <FlatButton
+                className="PendingResult-button"
+                onTouchTap={this.handleEditClick}
+              >
+                Edit
+              </FlatButton>}
           </CardActions>
         </Card>
       </div>
