@@ -1,6 +1,7 @@
-import React, { PropTypes } from 'react';
+import React from 'react';
 import { Card, CardHeader, CardTitle, CardText } from 'material-ui/Card';
 import Divider from 'material-ui/Divider';
+import PropTypes from 'prop-types';
 
 const Result = ({
   comments,
@@ -9,7 +10,7 @@ const Result = ({
   performanceName,
   spotId,
   winner
-}) => (
+}) =>
   <Card>
     <CardTitle title={performanceName} subtitle={performanceDate} />
     <Divider />
@@ -18,9 +19,10 @@ const Result = ({
       subtitle={winner ? 'You Won!' : 'You Lost'}
       avatar={winner ? '/images/trophy.png' : '/images/crying-cat-face.png'}
     />
-    <CardText>{comments}</CardText>
-  </Card>
-);
+    <CardText>
+      {comments}
+    </CardText>
+  </Card>;
 
 Result.propTypes = {
   comments: PropTypes.string.isRequired,
