@@ -1,9 +1,23 @@
 import api from './api';
 import auth from './auth';
+import changeCase from './change_case';
 import compareSpots from './compare-spots';
 import errorEmitter from './error-emitter';
-import identityFunction from './identity-function';
+import fetch from './fetch';
+import isEmptyObject from './is_empty_object';
+import noOp from './no_op';
 import * as routes from './routes';
 import screenSizes from './screen-sizes';
 
-export { api, auth, compareSpots, errorEmitter, identityFunction, routes, screenSizes };
+export {
+  api,
+  auth,
+  changeCase,
+  compareSpots,
+  errorEmitter,
+  fetch,
+  isEmptyObject,
+  noOp,
+  routes,
+  screenSizes
+};
